Update windows-build-release.yml workflow
This commit updates the windows-build-release.yml workflow file. It includes changes to the on event triggers and adds a new step to lookup the conclusion of the workflow_run event. The guard job has also been modified to check for a successful workflow_run conclusion and a branch starting with 'refs/tags/v'.